requirements.txt
asgiref==3.4.1
Django==3.2.12
django-filter==21.1
django-phonenumber-field==6.0.0
djangorestframework==3.13.1
phonenumberslite==8.12.44
pytz==2021.3
sqlparse==0.4.2
typing_extensions==4.1.1
